requirements.txt
Django==3.1.1
drf-yasg==1.17.1
djangorestframework==3.11.1
django-filter==2.3.0
django-debug-toolbar===2.2
